Don't crash if commit ref doesn't exist in upstream
When the script is collecting the set of included upstream commits,
if it doesn't find the referenced commit in upstream, it crashes.
This can be confusing for users and inefficient if multiple bad commits
exist.

Instead, catch Exceptions raised by GitPython if bad commit hash is
passed, print error to user, then continue processing.

ValueError is raised if the commit hash was not found in the repo, while
BadName is raised if the input is not a well-formed commit hash.

Signed-off-by: Dennis Chen <dechen@redhat.com>